+++ save-api/src/main/kotlin/org/cqfn/save/api/PropertiesConfiguration.kt
@@ -1,0 +1,86 @@
+package org.cqfn.save.api
+
+import org.slf4j.LoggerFactory
+import java.io.IOException
+import java.util.*
+
+private val log = LoggerFactory.getLogger(PropertiesConfiguration::class.java)
+
+enum class PropertiesConfigurationType {
+    EVALUATED_TOOL,
+    WEB_CLIENT,
+    ;
+}
+
+sealed class PropertiesConfiguration
+
+/**
+ * @property backendUrl
+ * @property preprocessorUrl
+ */
+data class WebClientProperties(
+    val backendUrl: String,
+    val preprocessorUrl: String,
+    val fileStorage: String,
+) : PropertiesConfiguration()
+
+/**
+ * @property organizationName
+ * @property projectName
+ * @property gitUrl
+ * @property gitUserName
+ * @property gitPassword
+ * @property branch
+ * @property commitHash
+ * @property testRootPath
+ */
+data class EvaluatedToolProperties(
+    val organizationName: String,
+    val projectName: String,
+    val gitUrl: String,
+    val gitUserName: String? = null,
+    val gitPassword: String? = null,
+    val branch: String? = null,
+    val commitHash: String?,
+    val testRootPath: String,
+    val additionalFiles: String? = null,
+) : PropertiesConfiguration()
+
+
+fun readPropertiesFile(configFileName: String, type: PropertiesConfigurationType): PropertiesConfiguration? {
+    try {
+        val properties = Properties()
+        val classLoader = AutomaticTestInitializator::class.java.classLoader
+        val input = classLoader.getResourceAsStream(configFileName)
+        if (input == null) {
+            log.error("Unable to find configuration file: $configFileName")
+            return null
+        }
+        properties.load(input)
+        when (type) {
+            PropertiesConfigurationType.WEB_CLIENT -> return WebClientProperties(
+                properties.getProperty("backendUrl"),
+                properties.getProperty("preprocessorUrl"),
+                properties.getProperty("fileStorage"),
+            )
+            PropertiesConfigurationType.EVALUATED_TOOL -> return EvaluatedToolProperties(
+                properties.getProperty("organizationName"),
+                properties.getProperty("projectName"),
+                properties.getProperty("gitUrl"),
+                properties.getProperty("gitUserName"),
+                properties.getProperty("gitPassword"),
+                properties.getProperty("branch"),
+                properties.getProperty("commitHash"),
+                properties.getProperty("testRootPath"),
+                properties.getProperty("additionalFiles"),
+            )
+            else -> {
+                log.error("Type $type for properties configuration doesn't supported!")
+                return null
+            }
+        }
+    } catch (ex: IOException) {
+        ex.printStackTrace()
+        return null
+    }
+}
